AB8828 - Created DatabaseBuilderTests
