+++ src/Umbraco.Infrastructure/Persistence/IEmbeddedDatabaseCreator.cs
@@ -3,6 +3,7 @@
     public interface IEmbeddedDatabaseCreator
     {
         string ProviderName { get; }
+        string ConnectionString { get; set; }
         void Create();
     }
 }
