+++ src/Umbraco.Tests.Integration/Umbraco.Infrastructure/Persistence/DatabaseBuilderTests.cs
@@ -1,0 +1,78 @@
+﻿using System;
+using System.IO;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
+using Moq;
+using NPoco;
+using NUnit.Framework;
+using Umbraco.Core;
+using Umbraco.Core.Configuration;
+using Umbraco.Core.Migrations.Install;
+using Umbraco.Core.Persistence;
+using Umbraco.Tests.Integration.Testing;
+using Umbraco.Tests.Testing;
+
+namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Persistence
+{
+    [TestFixture]
+    [UmbracoTest]
+    [Platform("Win")]
+    public class DatabaseBuilderTests : UmbracoIntegrationTest
+    {
+        private IDbProviderFactoryCreator DbProviderFactoryCreator => GetRequiredService<IDbProviderFactoryCreator>();
+        private IUmbracoDatabaseFactory UmbracoDatabaseFactory => GetRequiredService<IUmbracoDatabaseFactory>();
+        private IEmbeddedDatabaseCreator EmbeddedDatabaseCreator => GetRequiredService<IEmbeddedDatabaseCreator>();
+
+        [Test]
+        public void CreateDatabase()
+        {
+            var path = TestContext.CurrentContext.TestDirectory.Split("bin")[0];
+            AppDomain.CurrentDomain.SetData("DataDirectory", path);
+            const string dbFile = "DatabaseContextTests.sdf";
+            // delete database file
+            // NOTE: using a custom db file for this test since we're re-using the one created with BaseDatabaseFactoryTest
+            var filePath = string.Concat(path, dbFile);
+            if (File.Exists(filePath))
+                File.Delete(filePath);
+
+            EmbeddedDatabaseCreator.ConnectionString = $"Datasource=|DataDirectory|{dbFile};Flush Interval=1";
+
+            UmbracoDatabaseFactory.Configure(EmbeddedDatabaseCreator.ConnectionString, Constants.DbProviderNames.SqlCe);
+            DbProviderFactoryCreator.CreateDatabase(Constants.DbProviderNames.SqlCe);
+            UmbracoDatabaseFactory.CreateDatabase();
+
+            // test get database type (requires an actual database)
+            using (var database = UmbracoDatabaseFactory.CreateDatabase())
+            {
+                var databaseType = database.DatabaseType;
+                Assert.AreEqual(DatabaseType.SQLCe, databaseType);
+            }
+
+            // create application context
+            //var appCtx = new ApplicationContext(
+            //    _databaseFactory,
+            //    new ServiceContext(migrationEntryService: Mock.Of<IMigrationEntryService>()),
+            //    CacheHelper.CreateDisabledCacheHelper(),
+            //    new ProfilingLogger(Mock.Of<ILogger>(), Mock.Of<IProfiler>()));
+
+            // create the umbraco database
+            DatabaseSchemaCreator schemaHelper;
+            using (var database = UmbracoDatabaseFactory.CreateDatabase())
+            using (var transaction = database.GetTransaction())
+            {
+                schemaHelper = new DatabaseSchemaCreator(database, Mock.Of<ILogger<DatabaseSchemaCreator>>(), NullLoggerFactory.Instance, new UmbracoVersion());
+                schemaHelper.InitializeDatabaseSchema();
+                transaction.Complete();
+            }
+
+            var umbracoNodeTable = schemaHelper.TableExists("umbracoNode");
+            var umbracoUserTable = schemaHelper.TableExists("umbracoUser");
+            var cmsTagsTable = schemaHelper.TableExists("cmsTags");
+
+            Assert.That(umbracoNodeTable, Is.True);
+            Assert.That(umbracoUserTable, Is.True);
+            Assert.That(cmsTagsTable, Is.True);
+        }
+
+    }
+}
